:sparkles: Enhance game graphics by adding laser handling, updating display methods, and adjusting position calculations for coins and lasers

+++ src/shared/include/Obstacle.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 typedef struct obstacle_s {
-    int x_pos;
-    int y_pos;
+    double x_pos;
+    double y_pos;
 
     obstacle_s(int x = 0, int y = 0) : x_pos(x), y_pos(y) {}
 } obstacle_t;
